Simplify the predicates to remove-if in next-f.
In the case of adding new edges to f, we can first call adjacent-edges
to get a list of just the edges adjacent to v, and then filter out those
whose terminus is in pst. In the case of removing edges from f, there is
no need to call contains-vertex, since it will always return
true (because by design, f is a list of edges from verticies in pst to
vertices not in pst).